+++ v1_13_R2/src/main/java/be/isach/ultracosmetics/v1_13_R2/ridable/AIFindNearestPlayer.java
@@ -1,26 +1,16 @@
 package be.isach.ultracosmetics.v1_13_R2.ridable;
 
-import net.minecraft.server.v1_13_R2.AttributeInstance;
-import net.minecraft.server.v1_13_R2.Entity;
-import net.minecraft.server.v1_13_R2.EntityCreature;
-import net.minecraft.server.v1_13_R2.EntityHuman;
-import net.minecraft.server.v1_13_R2.EntityInsentient;
-import net.minecraft.server.v1_13_R2.EntityLiving;
-import net.minecraft.server.v1_13_R2.EntityPlayer;
-import net.minecraft.server.v1_13_R2.GenericAttributes;
-import net.minecraft.server.v1_13_R2.PathfinderGoal;
-import net.minecraft.server.v1_13_R2.PathfinderGoalNearestAttackableTarget;
-import net.minecraft.server.v1_13_R2.PathfinderGoalTarget;
-import net.minecraft.server.v1_13_R2.ScoreboardTeamBase;
+import net.minecraft.server.v1_13_R2.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bukkit.event.entity.EntityTargetEvent;
 
 import java.util.List;
 import java.util.function.Predicate;
+
 /**
  * @author BillyGalbreath
- *
+ * <p>
  * Author of plugin: "Ridables"
  * Thanks for authorizing using Ridables code to make UC work!
  */
@@ -47,14 +37,14 @@ public class AIFindNearestPlayer extends PathfinderGoal {
             } else {
                 double range = maxTargetRange();
                 if (target.isSneaking()) {
-                    range *= (double) 0.8F;
+                    range *= 0.8F;
                 }
                 if (target.isInvisible()) {
                     float f = ((EntityHuman) target).dk(); // getArmorVisibility
                     if (f < 0.1F) {
                         f = 0.1F;
                     }
-                    range *= (double) (0.7F * f);
+                    range *= 0.7F * f;
                 }
                 return (double) target.g(entity) <= range && PathfinderGoalTarget.a(entity, (EntityLiving) target, false, true);
             }
